Fix build of 11.3-019 on buster.

+++ main/base/os/daqshm.cpp
@@ -247,7 +247,7 @@ CDAQShm::attach(std::string name)
   
   // Add a new mapping entry with refcount 1:
 
-  attachInformation initialInfo = {pMemory, fileSize, 1};
+  attachInformation initialInfo = {pMemory, (size_t)fileSize, 1};
   m_attachMap[name] = initialInfo;
 
   return pMemory;
